[starts #161774149] restructure v1 test folder and write test for single order retrieval
